+++ Exodus/src/utopia/exodus/database/factory/description/DescriptionLinkFactory.scala
@@ -72,7 +72,7 @@ object DescriptionLinkFactory
 	/**
 	  * Role description factory
 	  */
-	lazy val role = apply(DescriptionLinkModel.role)
+	lazy val userRole = apply(DescriptionLinkModel.userRole)
 	
 	/**
 	  * Task description factory