+++ Metropolis/src/utopia/metropolis/model/partial/organization/MembershipData.scala
@@ -2,6 +2,20 @@ package utopia.metropolis.model.partial.organization
 
 import java.time.Instant
 
+import utopia.flow.datastructure.immutable.{Constant, Model, ModelDeclaration}
+import utopia.flow.generic.{FromModelFactoryWithSchema, InstantType, IntType, ModelConvertible}
+import utopia.flow.generic.ValueConversions._
+import utopia.flow.generic.ValueUnwraps._
+
+object MembershipData extends FromModelFactoryWithSchema[MembershipData]
+{
+	override val schema = ModelDeclaration("organization_id" -> IntType, "user_id" -> IntType,
+		"started" -> InstantType)
+	
+	override protected def fromValidatedModel(model: Model[Constant]) = MembershipData(model("organization_id"),
+		model("user_id"), model("inviter_id"), model("started"), model("ended"))
+}
+
 /**
   * Contains basic data about an organization membership
   * @author Mikko Hilpinen
@@ -13,4 +27,8 @@ import java.time.Instant
   * @param ended Timestamp when the membership ended. None if not ended (default).
   */
 case class MembershipData(organizationId: Int, userId: Int, creatorId: Option[Int] = None,
-						  started: Instant = Instant.now(), ended: Option[Instant] = None)
+						  started: Instant = Instant.now(), ended: Option[Instant] = None) extends ModelConvertible
+{
+	override def toModel = Model(Vector("organization_id" -> organizationId, "user_id" -> userId,
+		"inviter_id" -> creatorId, "started" -> started, "ended" -> ended))
+}